Started the custom trip view.

+++ app/src/main/java/com/andrewkingmarshall/andrewmarshalltripfileschallenge/Objects/Location.java
@@ -10,27 +10,27 @@ import java.io.Serializable;
 public class Location implements Serializable {
 
     @SerializedName("Latitude")
-    private long latitude;
+    private double latitude;
 
     @SerializedName("Longitude")
-    private long longitude;
+    private double longitude;
 
     @SerializedName("GpPlaceId")
     private String gpPlaceId;
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
